Added deep maze objective contents
