+++ levels/bias/objectives/objective2_5_deepmaze_1/description.md
@@ -4,16 +4,16 @@
 <h3>To-Do List</h3>
 <ul>
   <li>Watch the video.</li>
-  <li>Answer the question on the right.</li>
+  <li>Answer the true or false question.</li>
   <li>Click <em>HACK</em> to check your answer.</li>
 </ul>
 </div>
 
+Navigate throughout the **Stories Gallery** on the Cryo Bridge. Though each station is equally important, you'll need to complete at least **FOUR** stations to exit the gallery.
 
+There are **five** computer terminal stations in the gallery. After completing at least **FOUR** stations, go to the final gate on the top of the gallery to unlock the exit to the Cryo Bridge.
 
-## Watch the Video
-Watch the **video** below and see how well you fare with **(topic)**.
+Watch this video: 
+https://www.youtube.com/watch?v=PUHop5i8-f4
 
-link
-
-<iframe width="560" height="315" src="https://www.youtube-nocookie.com/embed/[link]" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+<iframe width="560" height="315" src="https://www.youtube-nocookie.com/embed/PUHop5i8-f4" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
