+++ levels/bias/objectives/objective2_5_deepmaze_1/walkthrough.md
@@ -1,3 +1,7 @@
 # Think Critically!
 
-help
+Answer the TRUE or False question to successfully complete the challenge at this story gallery station. 
+
+Remember: there are five computer terminal stations in the Stories Gallery on the Cryo Bridge.  After completing at least four stations, go to the final gate on the top of the gallery to unlock the exit to the Cryo Bridge.
+
+If you are unable to watch the video in the game, you can click on the link above the video to watch it in another browser.
