+++ levels/bias/objectives/objective2_5_deepmaze_2/objective.json
@@ -1,12 +1,12 @@
 {
-  "title": "Deepening Understanding Through Stories #2: (Topic)",
-  "description": "description",
+  "title": "Stories Gallery: Equality Trailblazers",
+  "description": "View these stories on marginalized folks and how bias in our world impacts them.",
   "validation_fields": [
     {
       "name": "answer1",
-      "type": "textarea",
+      "type": "boolean",
       "placeholder": "",
-      "label": "What do you think about this question?"
+      "label": "True or False: The companies with the most gender diverse executive teams were 21% more likely to outperform on profitability."
     }
   ],
   "rewards": {
