+++ levels/bias/objectives/objective2_5_deepmaze_2/validator.js
@@ -1,21 +1,6 @@
+const { processAnswer } = require("../lib/deepmaze_helper.js");
+
 module.exports = async function (helper) {
-    const worldState = helper.world.getState("com.twilioquest.Bias");
-    const deepMaze = worldState.Bias.deepMaze;
-
-    if (!worldState.Bias.objective.hasEnoughTimePassed) {
-        return helper.fail(`Did you watch the video? Make sure you watch the video for a while to really understand its topics.`);
-    }
-    
-    const {
-        answer1
-    } = helper.validationFields;
-
-    if (answer1.length < 100) {
-        return helper.fail(`
-            Try to write more and express your thoughts on the topic in the video.
-        `)
-    }
-    helper.success(`
-        "Thanks for your thoughts!"
-    `)
+    const correctAnswer = true;
+    return processAnswer(helper, correctAnswer);
 };
