+++ levels/bias/objectives/objective2_5_deepmaze_3/objective.json
@@ -1,12 +1,12 @@
 {
-  "title": "Deepening Understanding Through Stories #2: (Topic)",
-  "description": "description",
+  "title": "Stories Gallery: Hearing Hands",
+  "description": "View these stories on marginalized folks and how bias in our world impacts them.",
   "validation_fields": [
     {
       "name": "answer1",
-      "type": "textarea",
+      "type": "boolean",
       "placeholder": "",
-      "label": "What do you think about this question?"
+      "label": "True or False: The title and theme of the video “Widen the Screen” is asking us to not limit our view based on what images and stories are given to us."
     }
   ],
   "rewards": {
