+++ levels/bias/objectives/objective2_5_deepmaze_4/description.md
@@ -4,16 +4,24 @@
 <h3>To-Do List</h3>
 <ul>
   <li>Watch the video.</li>
-  <li>Answer the question on the right.</li>
+  <li>Answer the true or false question.</li>
   <li>Click <em>HACK</em> to check your answer.</li>
 </ul>
 </div>
 
+Navigate throughout the **Stories Gallery** on the Cryo Bridge. Though each station is equally important, you'll need to complete at least **FOUR** stations to exit the gallery.
+
+There are **five** computer terminal stations in the gallery. After completing at least **FOUR** stations, go to the final gate on the top of the gallery to unlock the exit to the Cryo Bridge.
+
+Watch these two videos: 
+
+1. Redrawing the Balance
+https://www.youtube.com/watch?v=qv8VZVP5csA
+
+<iframe width="560" height="315" src="https://www.youtube-nocookie.com/embed/qv8VZVP5csA" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+
+2. Queen Rules
+https://www.youtube.com/watch?v=Jcy2j5Mq8Tk
 
 
-## Watch the Video
-Watch the **video** below and see how well you fare with **(topic)**.
-
-link
-
-<iframe width="560" height="315" src="https://www.youtube-nocookie.com/embed/[link]" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+<iframe width="560" height="315" src="https://www.youtube-nocookie.com/embed/Jcy2j5Mq8Tk" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
