+++ levels/bias/objectives/objective2_5_deepmaze_5/objective.json
@@ -1,12 +1,18 @@
 {
-  "title": "Deepening Understanding Through Stories #2: (Topic)",
-  "description": "description",
+  "title": "What Dark-Skinned People Will Never Tell You",
+  "description": "View these stories on marginalized folks and how bias in our world impacts them.",
   "validation_fields": [
     {
       "name": "answer1",
-      "type": "textarea",
+      "type": "boolean",
       "placeholder": "",
-      "label": "What do you think about this question?"
+      "label": "True or False: Eurocentric standards of beauty impact people all over the world."
+    },
+    {
+      "name": "answer2",
+      "type": "boolean",
+      "placeholder": "",
+      "label": "True or False: One way we can combat colorism is to have more positive representations of dark skinned folks in the media."
     }
   ],
   "rewards": {
